+++ samples/s07-multipleGeometry/hostCode.cpp
@@ -73,8 +73,8 @@ template <typename T> struct Mesh {
     indexBuffer = gprtDeviceBufferCreate<uint3>(context, indices.size(), indices.data());
     geometry = gprtGeomCreate(context, geomType);
     TrianglesGeomData *geomData = gprtGeomGetParameters(geometry);
-    gprtTrianglesSetVertices(geometry, vertexBuffer, vertices.size());
-    gprtTrianglesSetIndices(geometry, indexBuffer, indices.size());
+    gprtTrianglesSetVertices(geometry, vertexBuffer, (uint32_t)vertices.size());
+    gprtTrianglesSetIndices(geometry, indexBuffer, (uint32_t)indices.size());
     geomData->vertex = gprtBufferGetHandle(vertexBuffer);
     geomData->index = gprtBufferGetHandle(indexBuffer);
     geomData->color = color;
@@ -202,10 +202,10 @@ main(int ac, char **av) {
 #endif
 
       // step 1 : Calculate the amount of rotation given the mouse movement.
-      float deltaAngleX = (2 * M_PI / fbSize.x);
-      float deltaAngleY = (M_PI / fbSize.y);
-      float xAngle = (lastxpos - xpos) * deltaAngleX;
-      float yAngle = (lastypos - ypos) * deltaAngleY;
+      float deltaAngleX = float(2 * M_PI / fbSize.x);
+      float deltaAngleY = float(M_PI / fbSize.y);
+      float xAngle = float(lastxpos - xpos) * deltaAngleX;
+      float yAngle = float(lastypos - ypos) * deltaAngleY;
 
       // step 2: Rotate the camera around the pivot point on the first axis.
       float4x4 rotationMatrixX = rotation_matrix(rotation_quat(lookUp, xAngle));
